This is for issue #1712
Simplify IBMi class by creating new IBMiSettings class that now holds
all the logic to check on IBM i settings and leaves IBMi class with the
connection logic.
